+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- Updated the layout of modal component
+- Added props containerClass and paddingContent to modal component to allow custom styling
+
 ### Added
 
 ### Fixed